Add files via upload
Space themed sequel to my first game. It is programmed completely in swift to be used on a variety of iOS devices. Here the player tries to destroy as many enemies as possible before getting hit by the enemy ultimately bringing an end to the game. Written in Swift Using xCode.